Correct issue with 4 element vectors being parsed as 3
The rotation matrix uses a three element vector and RPR creates 4 element vectors. This causes an error when trig is performed on the data.
In order to prevent it from erroring when default values are used (which only have three elements), I have made it slice the incoming rotation vector down to three elements.

If someone has a better way of doing this, such as if theres a macro for xyzw->xyz, please implement it. If not, this fixes any glaring issues for the time being.